+++ pkg/util/grpc/grpc.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -30,6 +30,7 @@ import (
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/metadata"
+	"google.golang.org/protobuf/encoding/protojson"
 	proto "google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/known/anypb"
 
@@ -143,4 +144,28 @@ func GRPCAnyCastAs[T proto.Message](in *anypb.Any, v T) error {
 	}
 
 	return nil
+}
+
+func NewGRPC[T proto.Message](in T) GRPC[T] {
+	return GRPC[T]{
+		Object: in,
+	}
+}
+
+type GRPC[T proto.Message] struct {
+	Object T
+}
+
+func (g *GRPC[T]) UnmarshalJSON(data []byte, opts ...util.Mod[protojson.UnmarshalOptions]) error {
+	o, err := Unmarshal[T](data, opts...)
+	if err != nil {
+		return err
+	}
+
+	g.Object = o
+	return nil
+}
+
+func (g GRPC[T]) MarshalJSON(opts ...util.Mod[protojson.MarshalOptions]) ([]byte, error) {
+	return Marshal[T](g.Object, opts...)
 }